// api/src/modules/local/dto/atualizar-local.dto.ts
export class AtualizarLocalDto {
    nome?: string;
    descricao?: string;
    endereco_logradouro?: string;
    endereco_numero?: string;
    endereco_complemento?: string;
    endereco_bairro?: string;
    endereco_cidade?: string;
    endereco_estado?: string;
    endereco_cep?: string;
    endereco_pais?: string;
    latitude?: number;
    longitude?: number;
    telefone_contato?: string;
    email_contato?: string;
    site?: string;
    amenidades?: number[]; // IDs das amenidades
}